+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package camp.nextstep;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
